+++ src/api/index.js
@@ -1,5 +1,5 @@
 export async function login({ email, imageUrl, name }) {
-  const request = await fetch(`${process.env.REACT_APP_SERVER_PORT}/api/login`, { 
+  const response = await fetch(`${process.env.REACT_APP_SERVER_PORT}/api/login`, { 
     method: "POST",
     headers: {
       "Accept": "application/json",
@@ -12,7 +12,7 @@ export async function login({ email, imageUrl, name }) {
     }),
   });
 
-  const { result, token } = request.json();
+  const { result, data: { token, user } } = await response.json();
 
-  return { result, token };
+  return { user, token };
 }
