+++ src/pages/LoginPage.js
@@ -1,0 +1,7 @@
+import LoginButton from "../components/LoginButton";
+
+export default function LoginPage({ onSuccess, onFailure }) {  
+  return (
+    <LoginButton onSuccess={onSuccess} onFailure={onFailure} />
+  );
+}
